requirements.txt
streamlit==1.42.0
pandas==2.1.4
numpy==1.26.4
plotly==5.22.0
bcrypt==4.1.2
Pillow==10.3.0
matplotlib==3.8.4
reportlab==4.0.9
